fix: advanced inference recommendation jobs parameters check
